requirements.txt
pyyaml
pyinstaller
paramiko
tqdm